Adding Login Register

+++ domain/src/main/java/com/deeromptech/domain/usecase/DeleteProductUseCase.kt
@@ -4,5 +4,5 @@ import com.deeromptech.domain.model.CartItemModel
 import com.deeromptech.domain.repository.CartRepository
 
 class DeleteProductUseCase(private val cartRepository: CartRepository) {
-    suspend fun execute(cartItemId: Int, userId: Int) = cartRepository.deleteItem(cartItemId, userId)
+    suspend fun execute(cartItemId: Int, userId: Long) = cartRepository.deleteItem(cartItemId, userId)
 }
